CLOiSim-2.2.0
[Major Changes]

* Upgrade Unity editor version
-> 2020.3.3f1

* SDF related
  * Add parsing Lights and spherical coordinates elements in World element
    * Import and implement light element
  * Add default argument for name and type attribute in SDF.entity
  * Separate Import.World method from Import.Loader
  * Gui element in World element is added for SDF parser
  * call ParseElements() method in SDF.Entity Base class
  * Apply rotation inertia when implement SDF link element
  * Do ImportModels() and ImportActors() methods in ImportWorlds() method

* Sensor Device related
  * Apply JobParallel computing for laser data  

  * Code refactoring for Device
    - Enhance Coroutine method
    - add Device Mode (TX/RX/None)
    - change OnVisualize() function abstract to virtual

  * Enhance Camera worker
    - do depth scaling buffer in CLOisim (moved from cloisim_ros)
    - remove 'request_realsense_paramter' interface
  
  * Adjust waiting period ratio for lidar

* Physics related
  * Cahange default properties for Root Articulation Body
    - set zero mass, minimum inertiatensor
  * Set 'Plane' Layer for plane collision object

* File and Folder Architectures
  * Separate the core scripts into Core directory
    - CLOiSimWorld
    - SphericalCoorditnates
    - SimulationService
    - BridgeManager

* Project Settings related
  * Add new tag
    - Light
  * Add new Layers
    - Plane


[Minor Changes]

* Versions
  * Update app version info in Project settings
    : 2.1.0 -> 2.2.0

  * Upgrade Plugin library
    - protobuf-net 3.0.73 -> 3.0.101

  * Update Packages
    - render-pipelines.core: 10.3.2 -> 10.4.0

* Code refactoring
  * Naming
    * Change method name in Clock class
      - calcualting simtime and realtime in Update() method
    * Change method name in Base Device class
      : MainDeviceWorker -> DeviceCoroutine
    * Change method name in SDF2Unity class
      : CheckTopModel() -> SDF2Unity.IsTopModel()
  * Code refactoring for SDF.Pose parser
    - support quaternion value for rotation
    - remove redundancy and enhance code reusability
  * Modify SDF2Unity
    - GetRotation() method -> converting quaternion values 
  * move 'relative_to' attribute in <pose> element into SDF.Pose class
  * Handling articulationbody in AfterImportModel() method
  * change the code to import SDF contents through Coroutine.

* Code cleanup
  * Remove SetArticulationBody() method in PoseControl
  * Remove useless code in SDF.World
  * Actors and Lights do not need to return object
  * Remove EnableSelfCollision() method since AriticulationBody is applied.
    - handling Joint Component in EnableSelfCollision() method is does not work anymore

* Logging related
  * Change Loglevel for assimp class
  * Suppress printing warning log in ParseElements() method.
  * Remove useless log print in GPS device

* Update Render setting in MainScene.unity

* Change Company name in Projectsettings


[Bug fix]

* PoseControl related
  * Reset the pose for only root articulation body in PoseControl

  * Enabling to reset the pose of Actors
    - reset waypoints
    - reset animation clip(rewind)

* GUI related
  * Prevent selecting the object if object has static property.
  * Prevent selecting the object which contain 'Plane Collision Object' for transform gizmo
  * Show only the object which has articulation body in Following object list

* Sensor Device related
  * Modify and fix GPS sensor deviee and spherical coordinates class
    - fix axis for spherical coordinates
    - Import reference values from world element for Spherical coordinates
  * Fix spherical coordinates and heading offset for GPS sensor device

* SDF related
  * Fix importing light
    - direction
  * fix error when world does not have actors
  * Fix wrong conversion in SphericalFromLocal() method
  * Add protection code for extracting resource paths from system environments
    - null check

+++ Assets/Scripts/Devices/Lidar.Filter.cs
@@ -32,8 +32,8 @@ namespace SensorDevices
 
 			if (numberOfBeams == laserScan.Intensities.Length)
 			{
-				var filterLowerBeamIndexRatio = (filterAngleLower_ - laserScan.AngleMin) / (laserScan.AngleMax - laserScan.AngleMin);
-				var filterUpperBeamIndexRatio = (filterAngleUpper_ - laserScan.AngleMin) / (laserScan.AngleMax - laserScan.AngleMin);
+				var filterLowerBeamIndexRatio = ((double)filterAngleLower_ - laserScan.AngleMin) / (laserScan.AngleMax - laserScan.AngleMin);
+				var filterUpperBeamIndexRatio = ((double)filterAngleUpper_ - laserScan.AngleMin) / (laserScan.AngleMax - laserScan.AngleMin);
 
 				filterLowerBeamIndex_ = (laserScan.AngleMin >= filterAngleLower_) ? (int?)null : (int)((double)numberOfBeams * filterLowerBeamIndexRatio);
 				filterUpperBeamIndex_ = (laserScan.AngleMax <= filterAngleUpper_) ? (int?)null : (int)((double)numberOfBeams * filterUpperBeamIndexRatio);
